(maint) Remove Puppet 3 compatiblity code
Previously the registry module was compatible with older Puppet agent versions
(< 4.7.0) however as the module only now supports 4.7.0 and above, much of this
code can be removed.  In particular many of the FFI calls and lower level API
methods are now contained within Puppet itself.  These should be used instead of
the custom written ones in this module to reduce code duplication and reduce
the testing burden.  This commit removes duplicate code, mainly from the base
class with some in the providers.  This commit also moves the valuename method
back into the registry_value provider as it is not used in the registry_key
provider so there is no need for it to be shared.