(MODULES-2957) Refactor path shared extraction
Previously the path extraction in the PuppetX::Puppetlabs::Registry conflated
validating paths and extracting valuenames from a path.  Only the
RegistryValuePath class required this functionality so it made no sense to be
shared in the RegistryKeyPath class.  The refactoring then assists in
MODULES-2957 which will add support for backslashes in the value name.

This commit:
* Moves the valuename extraction from the filter_path method into the initializer
  for the RegistryValuePath method.  This means that the path validation is
  actually happening on registry paths, not path + valuename.
* Now that the valuename was moved out, the code could be simplified.  This also
  meant the subkey method was now the same between then RegistryValuePath and
  RegistryKeyPath classes so it could be moved into the base class