(MODULES-2957) Add support for backslashes in registry value name
Previously it was not possible to modify a registry value with a backslash in
the name as the existing code used backslashes as the delimiter for registry
key paths.  As registry key names cannot have a backslash or be empty, the
delimiter between a registry key path and value is changed to a double backslash
for example:

HKLM\Software\Vendor\\ValueWithA\Backslash.

In this case the registry key path would be 'HKLM\Software\Vendor' and the value
name 'ValueWithA\Backslash'.

This commit:
* Uses a double backslash by default to determine the key path and valuename
  from a resource title, but will revert back to the old behaviour (single
  backslash) if none are found
* Adds tests to show the new functionality in the custom type (registry_value),
  the defined helper type (registry::value) and in the purging of registry_key
  resources